Split up `lib` and `tests` in dist.
Previously, the `lib` and `test` output was intermixed in `dist`.
This changes them to match the actual structure on disk (so there
is now a `dist/lib` and `dist/test`).

It also fixes the `files` blacklist of `!dist/tests` (which previously
was non-existent).

While working on this, I noticed that `dist` is not cleaned prior to
a build, leading to bizarre issues when files are removed/moved in
lib or test.  I added a `build:clean` script to handle clearing this
up.
